Add more specs and adjust some
